Ajout des moyens de contacts dans la description
